added program buttons to page
